Comply with 3.2.2 of SAML CORE 2.0-os.pdf
Don't include InResponseTo unless it is a response to a request.

We previously were requiring a saml_request_id. Now, if
'saml_request_id' is nil, remove the InResponseTo element.

This is to comply with 3.2.2 Complex Type StatusResponseType of the standard:
http://docs.oasis-open.org/security/saml/v2.0/saml-core-2.0-os.pdf

__InResponseTo__ [Optional]
A reference to the identifier of the request to which the response corresponds, if any. If the response
is not generated in response to a request, or if the ID attribute value of a request cannot be
determined (for example, the request is malformed), then this attribute MUST NOT be present.
Otherwise, it MUST be present and its value MUST match the value of the corresponding request's
ID attribute