Use FileUtils in suggested initializer (vs `mkdir`)
... for the creation of `./cache/saml/metadata` for the persisted metadata storage location.

Not only is the current implementation much more resource-intensive due to it spawning a new process on each invocation, it will fail should the rails root path contain spaces (or any other character that needs escaping). It also will quite obviously fail on platforms (that is to say, Windows) where the `mkdir` binary does not exist.