normalize Encryptor cert handling
This ensures the given `opts[:cert]` is formatted as expected in
the constructor. This removes the need for a type-check in the
`openssl_cert` method and DRYs up up removing any
`/-+(BEGIN|END) CERTIFICATE-+/` processing. This means you can
pass in `opts[:cert]` values in e.g. this form

```
-----BEGIN CERTIFICATE-----
ABC123...
-----END CERTIFICATE-----
```

and not get an `OpenSSL::X509::CertificateError nested asn1 error`
exception when calling `openssl_cert`.