Provide SAML request validation errors in payload
Previously, partners who sent an invalid SAML request would get a 403 and no
other helpful information. In this PR, the errors are returned as a JSON payload, and the status is a bad request instead of a 403.

I also fixed a typo in the error for `issuer_missing_or_invalid`.